backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
django-mongodb-backend==5.2.0b1
django-filter==24.3
PyJWT==2.9.0
python-dotenv==1.0.1
Pillow==10.4.0
python-dateutil==2.9.0
requests==2.32.3
celery==5.5.3
redis==5.0.8
boto3==1.34.144
django-extensions==3.2.3
cryptography==43.0.0
openpyxl==3.1.5
pandas==2.2.2
reportlab==4.0.4